Cleaned up REST inputs

+++ src/main/java/org/cytoscape/slimscape/internal/ui/QSlimfinderOptions.java
@@ -2,7 +2,7 @@ package org.cytoscape.slimscape.internal.ui;
 
 public class QSlimfinderOptions {
     private boolean dismask;
-    private double cutoff;
+    private int cutoff;
     private String length;
     private boolean consmask;
     boolean featuremask;
@@ -55,11 +55,11 @@ public class QSlimfinderOptions {
         this.wildcard = wildcard;
     }
 
-    public double getCutoff() {
+    public int getCutoff() {
         return cutoff;
     }
 
-    public void setCutoff(double cutoff) {
+    public void setCutoff(int cutoff) {
         this.cutoff = cutoff;
     }
